moved theme to its own component; made all style modular;
