+++ src/components/Face/index.js
@@ -1,5 +1,5 @@
 import { h, Component } from 'preact'
-import theme from '../../style/refactor.css'
+import theme from '../Theme/style.css'
 import style from './style.css'
 
 export const FaceTitle = ({ useCapture }) => {
