+++ src/components/Modal/style.css
@@ -1,4 +1,4 @@
-@import (less) "../../style/constants.css";
+@import (less) "../Theme/constants.css";
 
 .root {
   font-size: .875rem;
